Consolidate namespace structure
