+++ sources/Google.Solutions.IapDesktop.Extensions.Explorer.Test/ToolWindows/Install/TestAboutViewModel.cs
@@ -20,14 +20,13 @@
 //
 
 using Google.Solutions.IapDesktop.Application.Host;
-using Google.Solutions.IapDesktop.Application.Theme;
-using Google.Solutions.IapDesktop.Extensions.Explorer.Windows.About;
+using Google.Solutions.IapDesktop.Extensions.Explorer.ToolWindows.Install;
 using Google.Solutions.Testing.Application.Test;
 using Moq;
 using NUnit.Framework;
 using System;
 
-namespace Google.Solutions.IapDesktop.Extensions.Explorer.Test.Windows.About
+namespace Google.Solutions.IapDesktop.Extensions.Explorer.Test.ToolWindows.Install
 {
     [TestFixture]
     public class TestAboutViewModel : ApplicationFixtureBase
