+++ sources/Google.Solutions.IapDesktop.Extensions.Explorer/ToolWindows/Install/AboutView.resx
